requirements.txt
MetaTrader5==5.0.45
python-dotenv==1.0.0
numpy==1.24.3
pandas==2.0.3
ta-lib==0.4.28
pytz==2023.3
requests==2.31.0
